chore: move loadSharedConfigFiles to own folder
